Fixed linters error
