+++ webpack.config.js
@@ -29,6 +29,10 @@ module.exports = {
         test: /\.css/i,
         use: ['style-loader', 'css-loader'],
       },
+      {
+        test: /\.(png|svg|jpg|jpeg|gif)$/i,
+        type: 'asset/resource',
+      },
     ],
   },
 };
